requirements.txt
# Core dependencies
typer[all]>=0.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1

# API clients
google-generativeai>=0.3.0
elevenlabs>=0.2.0
httpx>=0.25.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
numpy>=1.24.0

# Utilities
rich>=13.7.0
python-dotenv>=1.0.0
tenacity>=8.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
mypy>=1.7.0
ruff>=0.1.0
black>=23.0.0